Registered Db Services in the PersistenceServiceregistration File

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -23,8 +24,8 @@ namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeave
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
             // verify that record exists
-            //if (leaveTypeToDelete == null)
-            //    throw new NotFoundException(nameof(LeaveType), request.Id);
+            if (leaveTypeToDelete == null)
+                throw new NotFoundException(nameof(LeaveType), request.Id);
 
             // remove from database
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
